% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{geolocart_locate}
\alias{geolocart_locate}
\title{Find locations in text and returns corresponding countries}
\usage{
geolocart_locate(text, geonameUser)
}
\arguments{
\item{text}{a string}

\item{geonameUser}{your geoname username}
}
\value{
vector of countries
}
\description{
Find locations in text and returns corresponding countries
}
\examples{
text <- "I live in Barcelona but I was born in Vannes and I have worked in Berlin in the past."
geolocart_locate(text, "insertYourName")
}

