% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{geolocart_locate}
\alias{geolocart_locate}
\title{Find locations in text}
\usage{
geolocart_locate(text)
}
\arguments{
\item{text}{a string}
}
\value{
vector of countries
}
\description{
Find locations in text
}
\examples{
text <- "I live in Barcelona but I was born in Vannes and I have worked in Berlin in the past."
geolocart_locate(text)
}

